+++ packages/plugin/tsdown.config.ts
@@ -6,12 +6,10 @@ import man from '../../manifest.json' with { type: 'json' };
 
 const dev = process.env.MODE === 'dev';
 const buildingPlugin = process.env.BUILD === 'plugin';
+const dtsPass = process.env.PASS === 'dts';
 
 const sharedConfig = defineConfig({
-	deps: {
-		neverBundle: ['obsidian'],
-		onlyBundle: false,
-	},
+	deps: { neverBundle: ['obsidian'], onlyBundle: false },
 	minify: true,
 	outExtensions: () => ({ dts: '.spec.d.ts', js: '.js' }),
 });
@@ -19,23 +17,14 @@ const sharedConfig = defineConfig({
 const pluginConfig = defineConfig({
 	...sharedConfig,
 	clean: !dev,
-	copy: [
-		{
-			from: '../../manifest.json',
-			to: 'dist-plugin',
-		},
-	],
+	copy: [{ from: '../../manifest.json', to: '../../dist' }],
 	css: {
 		fileName: 'styles.css',
 		minify: true,
-		postcss: {
-			plugins: [UnoCSS(), postcssMergeRules()],
-		},
+		postcss: { plugins: [UnoCSS(), postcssMergeRules()] },
 		transformer: 'postcss',
 	},
-	define: {
-		'Bun.env.VERSION': JSON.stringify(man.version),
-	},
+	define: { 'Bun.env.VERSION': JSON.stringify(man.version) },
 	dts: false,
 	entry: { main: 'src/index.ts' },
 	format: 'cjs',
@@ -46,10 +35,8 @@ const pluginConfig = defineConfig({
 			mainFields: ['browser', 'module', 'main'],
 		},
 	},
-	outDir: 'dist-plugin',
-	outputOptions: {
-		codeSplitting: false,
-	},
+	outDir: '../../dist',
+	outputOptions: { codeSplitting: false },
 	platform: 'browser',
 	// Fixes SolidJS cannot attach eventListeners to elements existing on another window
 	plugins: [solid({ solid: { delegateEvents: false } })],
@@ -58,13 +45,10 @@ const pluginConfig = defineConfig({
 
 const sdkConfig = defineConfig({
 	...sharedConfig,
-	clean: !dev,
-	dts: true,
-	entry: {
-		dev: 'src/sdk/dev.ts',
-		index: 'src/sdk/index.ts',
-	},
-	unbundle: true,
+	clean: !dev && dtsPass,
+	dts: dtsPass,
+	entry: { dev: 'src/sdk/dev.ts', index: 'src/sdk/index.ts' },
+	unbundle: !dtsPass,
 });
 
 export default buildingPlugin ? pluginConfig : sdkConfig;